[BUGFIX] Dev condition on platform.sh + branch name (#81)
Corrects the condition for development settings on Platform.sh. Also adds "DEV" and the current environment name to `$GLOBALS['TYPO3_CONF_VARS']['SYS']['sitename']`. For production, "PROD" is added. This helps differentiate environments. 

TYPO3_CONTEXT is not set (it wasn't before either). This should be done with an environment variable.